+++ packages/core/src/diff/enum.ts
@@ -41,11 +41,17 @@ export function changesInEnum(
 
     if (isNotEqual(oldValue.deprecationReason, newValue.deprecationReason)) {
       if (isVoid(oldValue.deprecationReason)) {
-        changes.push(enumValueDeprecationReasonAdded(newEnum, oldValue, newValue));
+        changes.push(
+          enumValueDeprecationReasonAdded(newEnum, oldValue, newValue),
+        );
       } else if (isVoid(newValue.deprecationReason)) {
-        changes.push(enumValueDeprecationReasonRemoved(newEnum, oldValue, newValue));
+        changes.push(
+          enumValueDeprecationReasonRemoved(newEnum, oldValue, newValue),
+        );
       } else {
-        changes.push(enumValueDeprecationReasonChanged(newEnum, oldValue, newValue));
+        changes.push(
+          enumValueDeprecationReasonChanged(newEnum, oldValue, newValue),
+        );
       }
     }
   });